chore: refactor and fix a hidden bug
+ All code and test code is formatted with black.
+ All code (but not test code) has been checked with pylint and flake8.
+ Removed a bug in `gradebook new` (see below for details).
+ Add pylint to requirements.txt.

Re the bug, I partially misunderstood how default arguments work in
Python functions.  I thought that if you had a function defined with
a default argument, and you passed in a variable with `None`, then the
default argument would be used.  This is wrong.  If you pass a variable
with the value `None`, then the argument gets the value `None`.  I fixed
the problem and added tests for the issue.