fix: remove race condition in `gradebook new`
The previous version of `gradebook new` had a race condition.  Briefly,
I would check whether a file named "whatever" existed before trying to
write a new file with that name.  This is a textbook race condition.

Happy, as of Python 3.3, you can open a file with "x" mode to get the
desired behavior.  This is equivalent to the lower-level O_EXCL mode.

I found this solution here: https://stackoverflow.com/a/18474773.