feat: improve feedback if user makes a bad request
A common error is that the user runs `gradebook calc -s3` when they mean
to run `gradebook calc -q3`.  Before this fix, the program would error
with an ugly stacktrace.

This update still exits in a failure mode, but now the user receives
a much clearer message about what they did wrong and how to fix their
mistake.